Breakup tests, move ConfigCollector
