+++ internal-api/src/main/java/datadog/trace/api/telemetry/ConfigCollector.java
@@ -1,5 +1,7 @@
-package datadog.trace.api;
+package datadog.trace.api.telemetry;
 
+import datadog.trace.api.ConfigOrigin;
+import datadog.trace.api.ConfigSetting;
 import java.util.Collections;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
